SW-45 * Fixed some setup and merged in develop
