+++ src/app/warehouse/warehouse-routing.module.ts
@@ -1,8 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomePageComponent } from './home-page/home-page.component';
+import { CreateInboundOrderComponent } from './create-inbound-order/create-inbound-order.component';
 
 const routes: Routes = [
+  { path: 'createInboundOrder', component: CreateInboundOrderComponent },
   { path: 'home', component: HomePageComponent },
   { path: '', redirectTo: 'home', pathMatch: 'full' }
 ];
